+++ src/components/Home.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const Home = () => {
+    return (
+        <div>
+            <h1>Home</h1>
+        </div>
+    )
+}
