+++ src/components/Home1.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const Home1 = () => {
+    return (
+        <div>
+            <h1>Home1</h1>
+        </div>
+    )
+}
