+++ src/components/Home2.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const Home2 = () => {
+    return (
+        <div>
+            <h1>Home2</h1>
+        </div>
+    )
+}
