+++ src/components/Home3.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const Home3 = () => {
+    return (
+        <div>
+            <h1>Home3</h1>
+        </div>
+    )
+}
